OC-441 da and microgrids entities RtuDevice and Task were entities for the same database table. Refactor to one class in core
