+++ osgp-adapter-domain-microgrids/src/main/java/org/opensmartgridplatform/adapter/domain/microgrids/application/tasks/CommunicationMonitoringTask.java
@@ -15,9 +15,9 @@ import java.util.List;
 import org.joda.time.DateTime;
 import org.opensmartgridplatform.adapter.domain.microgrids.application.services.CommunicationRecoveryService;
 import org.opensmartgridplatform.domain.core.valueobjects.DeviceLifecycleStatus;
-import org.opensmartgridplatform.domain.microgrids.entities.RtuDevice;
-import org.opensmartgridplatform.domain.microgrids.repositories.RtuDeviceRepository;
-import org.opensmartgridplatform.domain.microgrids.repositories.TaskRepository;
+import org.opensmartgridplatform.domain.core.entities.RtuDevice;
+import org.opensmartgridplatform.domain.core.repositories.RtuDeviceRepository;
+import org.opensmartgridplatform.domain.core.repositories.TaskRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
