+++ osgp-domain-core/src/main/java/org/opensmartgridplatform/domain/core/entities/Task.java
@@ -1,11 +1,11 @@
 /**
- * Copyright 2017 Smart Society Services B.V.
+ * Copyright 2016 Smart Society Services B.V.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License.  You may obtain a copy of the License at
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.opensmartgridplatform.domain.da.entities;
+package org.opensmartgridplatform.domain.core.entities;
 
 import java.util.Date;
 
@@ -14,7 +14,7 @@ import javax.persistence.Entity;
 import javax.persistence.EnumType;
 import javax.persistence.Enumerated;
 
-import org.opensmartgridplatform.domain.da.valueobjects.TaskStatusType;
+import org.opensmartgridplatform.domain.core.valueobjects.TaskStatusType;
 import org.opensmartgridplatform.shared.domain.entities.AbstractEntity;
 
 @Entity
