--- osgp-domain-distributionautomation/src/main/java/org/opensmartgridplatform/domain/da/repositories/TaskRepository.java
@@ -1,15 +1,0 @@
-/**
- * Copyright 2017 Smart Society Services B.V.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- */
-package org.opensmartgridplatform.domain.da.repositories;
-
-import org.opensmartgridplatform.domain.da.entities.Task;
-import org.springframework.data.repository.CrudRepository;
-
-public interface TaskRepository extends CrudRepository<Task, Long> {
-    Task findByTaskIdentification(String taskIdentification);
-}
